Refactor - move MoreCardsInHandThanOpponentsCondition to shared file
